Refactor to OOP format
Invalid commands do not seem to be caught by exception handler for some reason
