+++ build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'java'
     id 'application'
-    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '5.1.0'
 }
 
@@ -35,10 +34,6 @@ application {
 shadowJar {
     archiveBaseName = "duke"
     archiveClassifier = null
-}
-
-checkstyle {
-    toolVersion = '10.2'
 }
 
 run{
